gen: use msgp_ prefix on random variable names
Fix a case I was seeing where `var` was generated as a "unique" variable name.
Of course, the compiler complained about `var var`.

The issue wasn't easily reproduceable because it depends on `rand.Intn`, but
fortunately, msgp doesn't seed the random generator and I was able to reproduce
the issue enough times to find a fix.