Removed unnecessary "err" of naked return
The generated encode functions that have a naked return (simply "return") where the function returns an error type all have the "err" variable declared as a named return (as it is here in the func "appendraw"). But this is the only place where the "return" is followed by a " err"; this is unnecessary because the return value is named in the function signature. I've been manually removing the extra space-err in every .Append block because they make the code inconsistent and are very annoying.