Renamed gen.Printer and gen.applyall
The name "GeneratorSet" is much more suitable to what []generator does. Also, because func "applyall" was not capitalized right it was hard to read. Also (minor thing), this commit renames the FileSet method receiver names from "f" to "fs" where it was still "f" so that it's the same everywhere. This all will make the code more manageable.