Fix empty map case
An empty map in the data should result in an empty map in Go. It's a
small performance loss in the rare cases where people have empty maps in
their data and use structs with maps that haven't been initialized yet.
But for correctness its better and it allows people to distinguish
between no map in the data or an empty map.