Always create slices, even 0 length
This change makes it so that when a 0 length slice is present in the
msgp data a 0 length slice will be allocated in Go, instead of nil.

This means msgp data generated using this library will always result in
a slice being allocated as this library doesn't contain any conditional
code to not emit the slice header if a slice is nil. Other libraries
that do omit the slice header will still result in a nil slice if the
slice is completly missing from the msgp data.