Add native json.Number support
Allow encoding and decoding of `json.Number` values, either as struct members or as interface members.

Numbers will be encoded as integer, if possible, otherwise float64 is used. To cover the zero values, write an empty string, but invalid values will return errors when encoded

It is possible to encode as string with `//msgp:replace json.Number with:string`.

Fixes #292