Use float32 when lossless for float64
When a float64 can be fully represented as a float32, use that instead.

Encoding is backwards compatible with existing reader. `AppendFloat64Only` is made available to unconditionally write a float64.

Micro-benchmarks are not great:

```
BenchmarkAppendFloat64Only-32    	511061935	         2.324 ns/op	3442.55 MB/s	       0 B/op	       0 allocs/op
BenchmarkAppendFloat64-32    	136697270	         8.704 ns/op	 919.08 MB/s	       0 B/op	       0 allocs/op
```

But microbenchmarks are micro-benchmarks. We take a much bigger cost on `AppendInt64` and other integer functions.

We could make it opt-in, but that would make it pretty much unused.